Refactored my stats graph to separate component
