+++ frontend/src/components/hours-in-hubb-graph/HoursInHubbGraph.module.scss
@@ -1,11 +1,4 @@
-.userStatsCardsWrapper {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  margin: 2em 0;
-}
-
-.infoContainer {
+.graph {
   background-color: #fff;
   border: none;
   box-shadow: 0 2px 2px 0 rgba(0, 0, 0, 0.14), 0 3px 1px -2px rgba(0, 0, 0, 0.2), 0 1px 5px 0 rgba(0, 0, 0, 0.12);
@@ -19,13 +12,21 @@
   }
 }
 
-.infoHeader {
+.graphContainer {
+  background-color: #fff;
+  border: none;
+  box-shadow: 0 2px 2px 0 rgba(0, 0, 0, 0.14), 0 3px 1px -2px rgba(0, 0, 0, 0.2), 0 1px 5px 0 rgba(0, 0, 0, 0.12);
+  border-radius: 2px;
+  overflow: hidden;
+  flex-grow: 1;
+}
+
+.graphHeader {
   color: #ffffff;
   background-color: #00b3e6;
   padding: 0.5rem;
 }
 
-.infoText {
-  height: 4em;
-  padding: 0.4em;
+.graphContent {
+  padding: 10px 10px 10px 0;
 }
