+++ frontend/src/pages/users/[cid].tsx
@@ -6,6 +6,7 @@ import Head from 'next/head';
 
 import { MeCidQueryQuery, UserStatsQuery, UserStatsQueryVariables } from '../../__generated__/graphql';
 import Error from '../../components/error/Error';
+import HoursInHubbGraph, { USER_LAST_SESSIONS_FRAGMENT } from '../../components/hours-in-hubb-graph/HoursInHubbGraph';
 import UserStatsCards, { USER_STATS_FRAGMENT } from '../../components/user-stats-cards/UserStatsCards';
 import { createTitle, defaultGetServerSideProps, formatNick, PageProps } from '../../util';
 
@@ -15,10 +16,12 @@ const USER_STATS_QUERY = gql`
   query UserStats($input: UserUniqueInput!) {
     user(input: $input) {
       ...UserStats
+      ...UserLastSessions
     }
   }
 
   ${USER_STATS_FRAGMENT}
+  ${USER_LAST_SESSIONS_FRAGMENT}
 `;
 
 const ME_CID_QUERY = gql`
@@ -41,8 +44,9 @@ const UserStats: NextPage<PageProps<UserStatsQuery>> = ({ data }) => {
       </Head>
       <div className={styles.showSection}>
         <h1>{data.user.nick}</h1>
-        <div className={styles.showSectionF}>
+        <div className={styles.showSection}>
           <UserStatsCards user={data.user} />
+          <HoursInHubbGraph user={data.user} />
         </div>
       </div>
     </>
